Sorry, the last commit did not include the message by mistake.
These are the content changes from Larry Holish. There are also wording
changes here, since the hunks were big. Also it should be mentioned
that Larry's new description of what a honeypot is, and the reference
to Deception Toolkit, were committed in the previous patch... just wanted
to highlight the contributions.


git-svn-id: svn://svn.debian.org/ddp/manuals/trunk@1828 313b444b-1b9f-4f58-a734-7bb04f332e8d