../dynamic.ent not dynamic.ent is a .PHONY: target (this makefile is called from translation directories).
Makefile.common checked in the past only that dynamic_$(architecture).ent
exists otherwise it created this file and linked dynamic.ent to it.
But if dynamic_$(architecture).ent already exists and if dynamic.ent is a
proper link it will not update the link to refer to the new architecture.

The publish target wasn't affected since it performed a make clean first
so that the links where properly created.


git-svn-id: svn://svn.debian.org/ddp/manuals/trunk@2757 313b444b-1b9f-4f58-a734-7bb04f332e8d