Do no longer automatically update PO files if English files were changed
This is e.g. useful if the initial svn checkout is slow and the PO file is
older than .dbk files. The build should not update the PO files in the working
copy as this change could cause a conflict on the build host after the next svn
update. Also updating PO files with automatic tools does not affect the built
document as the translation statistics are not changed by this (only source
references will be updated, ...).

Only the target updatepo is now responsible to update PO files, independent of
timestamps. This closes: #509759.

Improved also the documentation a little bit.


git-svn-id: svn://svn.debian.org/ddp/manuals/trunk@5703 313b444b-1b9f-4f58-a734-7bb04f332e8d