Improve statistics by not counting PO file conflicts
A strings (such as the PO header) which is translated inconsistent across
multiple files would result in a PO conflict message which is fuzzy. This
leads e.g. to one bogus fuzzy message for Vietnamese.

Now only the first translation is used in case multiple are available.

This problem would not exist with a single PO file only but as we use
multiple ...


git-svn-id: svn://svn.debian.org/ddp/manuals/trunk@6599 313b444b-1b9f-4f58-a734-7bb04f332e8d