Fix TeX include path issues
Let's no longer copy ifxetex.sty into lang/ as the dblatex LaTeX build happens
in a temporary directory and the root directory of the Release Notes is already
in the TeX search path whereas lang/ isn't.

Also do not longer consider "es" special but add the language specific directory
to the default search path for all languages and also for the .ps target.


git-svn-id: svn://svn.debian.org/ddp/manuals/trunk@6600 313b444b-1b9f-4f58-a734-7bb04f332e8d